+++ src/main/java/nl/opengeogroep/safetymaps/utils/SafetyConnectMessageReceiver.java
@@ -286,8 +286,8 @@ public class SafetyConnectMessageReceiver implements ServletContextListener {
     try {
       // Is message for me
       if (unitIsForMyRegion(move, Arrays.asList(RQ_REGIONS.split(",")))) {
-        BigDecimal lon = (BigDecimal)move.get("lon");
-        BigDecimal lat = (BigDecimal)move.get("lat");
+        Double lon = (Double)move.get("lon");
+        Double lat = (Double)move.get("lat");
         Integer speed = move.has("speed") && move.get("speed").toString() != "null" ? move.getInt("speed") : 0;
         Integer heading = move.has("heading") && move.get("heading").toString() != "null" ? move.getInt("heading") : 0;
         Integer eta = move.has("eta") && move.get("eta").toString() != "null" ? move.getInt("eta") : 0;
